% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pa_get_event_data}
\alias{pa_get_event_data}
\title{Retrieve event data associated with a guid}
\usage{
pa_get_event_data(guid, last_time = 0)
}
\arguments{
\item{guid}{string: the match guid as returned by \code{\link{pa_get_matches}}}

\item{last_time}{datetime or numeric: only return events more recent than this}
}
\value{
Character.
}
\description{
Retrieve event data associated with a guid
}
