% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{pa_get_match_file}
\alias{pa_get_match_file}
\title{Create match file via API}
\usage{
pa_get_match_file(guid, filename = NULL)
}
\arguments{
\item{guid}{string: the match guid as returned by \code{\link{pa_get_matches}}}

\item{filename}{string: the filename to use. A temporary file will be used if no \code{filename} is provided. Existing files will not be overwritten}
}
\value{
Filename
}
\description{
Create match file via API
}
\examples{

\dontrun{
  my_server_name <- "server name used in VBStats"

  ## retrieve the ID of the first match in our collection
  guid <- pa_get_matches(my_server_name)$guid[1]

  ## build the match file
  fn <- pa_get_match_file(guid)

  ## load it
  x <- peranavolley::pv_read(fn)
}

}
