% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pa_get_match_info}
\alias{pa_get_match_info}
\title{Retrieve the match information associated with a guid}
\usage{
pa_get_match_info(guid)
}
\arguments{
\item{guid}{string: the match guid as returned by \code{\link{pa_get_sessions}}}
}
\value{
character
}
\description{
Retrieve the match information associated with a guid
}
