% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pa_get_matches}
\alias{pa_get_matches}
\title{Retrieve the matches associated with a given server name}
\usage{
pa_get_matches(server_name, app_name = "VBStats")
}
\arguments{
\item{server_name}{string: the server name used in VBStats}

\item{app_name}{string: the app name ("VBStats" for VBstatsHD)}
}
\value{
data.frame with columns \code{guid}, \code{teams}, \code{date}
}
\description{
Retrieve the matches associated with a given server name
}
\examples{
\dontrun{
 my_server_name <- "server name used in VBStats"

 ## retrieve the matches available to us
 pa_get_matches(my_server_name)

}

}
