% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pa_get_sessions}
\alias{pa_get_sessions}
\title{Retrieve the sessions associated with a given server name}
\usage{
pa_get_sessions(server_name, app_name = "VBStats")
}
\arguments{
\item{server_name}{string: the server name used in VBStats}

\item{app_name}{string: the app name ("VBStats" for VBstatsHD)}
}
\value{
data.frame with columns \code{guid}, \code{teams}, \code{date}
}
\description{
Retrieve the sessions associated with a given server name
}
